use i32 to avoid some integer casting

+++ src/c_interface.rs
@@ -578,9 +578,9 @@ macro_rules! c_accessor_array {
 
         pub fn $rust_index(
             &self,
-            index: usize,
+            index: i32,
         ) -> Option<crate::c_interface::CTmpRef<Self, $type>> {
-            if index < self.$count_fn() as usize {
+            if index < self.$count_fn() {
                 Some(crate::c_interface::CTmpRef::new(self, unsafe {
                     <$type as crate::c_interface::NewFromPtr<$c_type>>::new_from_ptr(
                         *self.c_ptr_ref().$c.offset(index as isize),
@@ -593,9 +593,9 @@ macro_rules! c_accessor_array {
 
         pub fn $rust_index_mut(
             &mut self,
-            index: usize,
+            index: i32,
         ) -> Option<crate::c_interface::CTmpMut<Self, $type>> {
-            if index < self.$count_fn() as usize {
+            if index < self.$count_fn() {
                 Some(crate::c_interface::CTmpMut::new(self, unsafe {
                     <$type as crate::c_interface::NewFromPtr<$c_type>>::new_from_ptr(
                         *self.c_ptr_mut().$c.offset(index as isize),
